Initial commit of the business parts!
Designed without the machine on hand, so there'll likely be a few
updates once she comes in, but I did leave some play for the drive
brackets.

Bed plates need to be updated with holes for the wires.

Need some standoffst for the bed plates.

The conveyor belt axles are situated 10mm above the rails, and are
12mm shafts - so there's 4mm clearance.  Bed should finish 16mm
above the rails, so would need 10mm standoffs to attach the bed
plates.

I'll work on a BOM in a bit... kinda wanna eyeball the bits first.