Multiple updates!
Updated the rear motor plate for larger 8mm flanged bearings.
Updated the rear camera mount - just a bit larger and angled down
towards the bed.
Updated the start and end gcodes - it'd been a while
Added a Pi 3 case for the rear mount system

Experimental:
Added an 80mm long motor plate standoff... in order to fit:
Spring Coupler.  Yeah.  The couplers are both springs.  The idea here is
that the spring compresses on homing, then springs back during ejection.
That way, the coupler can be in any orientation at all and still work;
the current one is reliable, but once it gets out of place it stays that
way.

Linear Advance calculation - only really needed for the Orbiter.

Thanks,
Paul